requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
exceptiongroup==1.1.3
fastapi==0.103.0
filelock==3.12.3
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.1
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
